Conditionally include scripts
