+++ app/javascript/packs/articlePage.jsx
@@ -1,11 +1,7 @@
-import { h, render } from 'preact';
-import { Snackbar, addSnackbarItem } from '../Snackbar';
 import { addFullScreenModeControl } from '../utilities/codeFullscreenModeSwitcher';
 import { initializeDropdown } from '../utilities/dropdownUtils';
 import { setupBillboardInteractivity } from '../utilities/billboardInteractivity';
 import { embedGists } from '../utilities/gist';
-import { initializeUserSubscriptionLiquidTagContent } from '../liquidTags/userSubscriptionLiquidTag';
-import { trackCommentClicks } from '@utilities/ahoy/trackEvents';
 import { isNativeAndroid, copyToClipboard } from '@utilities/runtime';
 
 // Open in new tab backfill
@@ -34,15 +30,6 @@ function backfillLinkTarget() {
   });
 }
 
-const animatedImages = document.querySelectorAll('[data-animated="true"]');
-if (animatedImages.length > 0) {
-  import('@utilities/animatedImageUtils').then(
-    ({ initializePausableAnimatedImages }) => {
-      initializePausableAnimatedImages(animatedImages);
-    },
-  );
-}
-
 const fullscreenActionElements = document.getElementsByClassName(
   'js-fullscreen-code-action',
 );
@@ -50,15 +37,6 @@ const fullscreenActionElements = document.getElementsByClassName(
 if (fullscreenActionElements) {
   addFullScreenModeControl(fullscreenActionElements);
 }
-
-// The Snackbar for the article page
-const snackZone = document.getElementById('snack-zone');
-if (snackZone) {
-  render(<Snackbar lifespan={3} />, snackZone);
-}
-
-// eslint-disable-next-line no-restricted-globals
-top.addSnackbarItem = addSnackbarItem;
 
 const multiReactionDrawerTrigger = document.getElementById(
   'reaction-drawer-trigger',
@@ -134,60 +112,10 @@ document
   .getElementById('copy-post-url-button')
   ?.addEventListener('click', copyArticleLink);
 
-// Comment Subscription
-getCsrfToken().then(async () => {
-  const { user = null, userStatus } = document.body.dataset;
-  const root = document.getElementById('comment-subscription');
-  const isLoggedIn = userStatus === 'logged-in';
-
-  if (!root) {
-    return;
-  }
-  try {
-    const {
-      getCommentSubscriptionStatus,
-      setCommentSubscriptionStatus,
-      CommentSubscription,
-    } = await import('../CommentSubscription');
-
-    const { articleId } = document.getElementById('article-body').dataset;
-
-    let subscriptionType = 'not_subscribed';
-
-    if (isLoggedIn && user !== null) {
-      ({ config: subscriptionType } = await getCommentSubscriptionStatus(
-        articleId,
-      ));
-    }
-
-    const subscriptionRequestHandler = async (type) => {
-      const message = await setCommentSubscriptionStatus(articleId, type);
-
-      addSnackbarItem({ message, addCloseButton: true });
-    };
-
-    render(
-      <CommentSubscription
-        subscriptionType={subscriptionType}
-        positionType="static"
-        onSubscribe={subscriptionRequestHandler}
-        onUnsubscribe={subscriptionRequestHandler}
-        isLoggedIn={isLoggedIn}
-      />,
-      root,
-    );
-  } catch (e) {
-    root.innerHTML =
-      '<p className="color-accent-danger">Unable to load Comment Subscription component.<br />Try refreshing the page.</p>';
-  }
-});
-
 const targetNode = document.querySelector('#comments');
 targetNode && embedGists(targetNode);
 
 setupBillboardInteractivity();
-initializeUserSubscriptionLiquidTagContent();
 focusOnComments();
 // Temporary Ahoy Stats for comment section clicks on controls
-trackCommentClicks('comments');
 backfillLinkTarget();
