+++ app/views/articles/show.html.erb
@@ -4,9 +4,13 @@
   .html-variant-wrapper { display: none}
 </style>
 
+<% if @article.processed_html.include?("data-animated") %>
+  <%= javascript_include_tag "articleAnimations", defer: true %>
+<% end %>
+
 <%= render "shared/webcomponents_loader_script" %>
 <% if user_signed_in? %>
-  <%= javascript_include_tag "webShare", "articlePage", "articleModerationTools", "commentDropdowns", defer: true %>
+  <%= javascript_include_tag "webShare", "articlePage", "articleSignedIn", "articleModerationTools", "commentDropdowns", defer: true %>
 <% else %>
   <%= javascript_include_tag "webShare", "articlePage", "commentDropdowns", defer: true %>
 <% end %>
